updated Cpython Linux to 23R1

+++ _unittest/conftest.py
@@ -26,6 +26,7 @@ import sys
 import tempfile
 import time
 
+from pyaedt import is_windows
 from pyaedt import pyaedt_logger
 from pyaedt import settings
 from pyaedt.generic.general_methods import generate_unique_name
@@ -123,10 +124,11 @@ class BasisTest(object):
             except:
                 pass
         if self.desktop:
-            try:
-                os.kill(self._main.desktop_pid, 9)
-            except:
-                pass
+            if is_windows:
+                try:
+                    os.kill(self._main.desktop_pid, 9)
+                except:
+                    pass
             self.desktop.release_desktop(False, True)
 
         del self.edbapps
